Move getSarcophagusState to Utils class
